+++ tests/binary_parser.cpp
@@ -15,8 +15,8 @@
  * @param hex The hexadecimal vector
  * @return The binary vector
  */
-std::vector<double> binary_from_hexadecimal(std::vector<uint8_t>&& hex) {
-	std::vector<double> result;
+std::vector<float> binary_from_hexadecimal(std::vector<uint8_t>&& hex) {
+	std::vector<float> result;
 	for (uint8_t byte : hex) {
 		for (int j = 7; j >= 0; j--) {
 			result.push_back((byte >> j) & 1);
@@ -31,8 +31,8 @@ std::vector<double> binary_from_hexadecimal(std::vector<uint8_t>&& hex) {
 TEST(BinaryWithSplit_Parser, Single_Instruction) {
 	std::string hexa = ";; Cycles: x\n\n"
 					   "FF FF\n";
-	std::vector<double> expected = binary_from_hexadecimal({0xFF, 0xFF});
-	std::vector<double> result = BinaryParser().parse_in(hexa);
+	std::vector<float> expected = binary_from_hexadecimal({0xFF, 0xFF});
+	std::vector<float> result = BinaryParser().parse_in(hexa);
 	EXPECT_EQ(result, expected);
 
 	hexa = ";; Cycles: x\n\n"
@@ -52,15 +52,15 @@ TEST(BinaryWithSplit_Parser, Multiple_Instructions) {
 							   "00\n"
 							   "71 D9\n";
 
-	std::vector<double> expected;
-	std::vector<double> inst1 = binary_from_hexadecimal({0xAA, 0xBB});
-	std::vector<double> inst2 = binary_from_hexadecimal({0xC5, 0xE4, 0xB8});
-	std::vector<double> inst3 = binary_from_hexadecimal({0x00});
-	std::vector<double> inst4 = binary_from_hexadecimal({0x71, 0xD9});
+	std::vector<float> expected;
+	std::vector<float> inst1 = binary_from_hexadecimal({0xAA, 0xBB});
+	std::vector<float> inst2 = binary_from_hexadecimal({0xC5, 0xE4, 0xB8});
+	std::vector<float> inst3 = binary_from_hexadecimal({0x00});
+	std::vector<float> inst4 = binary_from_hexadecimal({0x71, 0xD9});
 	expected.insert(expected.end(), inst1.begin(), inst1.end());
 	expected.insert(expected.end(), inst2.begin(), inst2.end());
 	expected.insert(expected.end(), inst3.begin(), inst3.end());
 	expected.insert(expected.end(), inst4.begin(), inst4.end());
-	std::vector<double> result = BinaryParser().parse_in(instructions);
+	std::vector<float> result = BinaryParser().parse_in(instructions);
 	EXPECT_EQ(result, expected);
 }
